Update README.md
In this project we perform linear regressions on a real dataset from a European Toyota
car dealer on the sales records of used cars (Toyota Corolla). We would like to construct a
reasonable linear regression model for the relationship between the sales prices of used cars
and various explanatory variables (such as age, mileage, horsepower). We are interested to see
what factors affect the sales price of a used car and by how much.